Update modbusTCP.cs
I created for test a project with a client and a server. I tried to communicate with them through the .dot net project and see what happened when the client and server program were shuttled down. It turned off before the "Exception data event" was caught. With the proposed change it works properly.